Move bookingSearch casting utilities
