+++ server/utils/bookingSearchUtils.ts
@@ -1,5 +1,10 @@
 import type { BookingSearchApiStatus } from '@approved-premises/ui'
-import { BookingSearchSortField } from '@approved-premises/api'
+import {
+  BookingSearchResult,
+  BookingSearchSortField,
+  Cas3BookingSearchResult,
+  Cas3BookingStatus,
+} from '@approved-premises/api'
 import { SubNavObj, TableCell } from '../@types/ui/index'
 import paths from '../paths/temporary-accommodation/manage'
 import { appendQueryString } from './utils'
@@ -47,3 +52,27 @@ export function capitaliseStatus(status: BookingSearchApiStatus) {
   const uiStatus = convertApiStatusToUiStatus(status)
   return uiStatus.charAt(0).toUpperCase() + uiStatus.slice(1)
 }
+
+// TODO -- ENABLE_CAS3V2_API cleanup: remove the following casting utilities and all usages
+export const isCas3BookingSearchResults = (
+  results: Array<BookingSearchResult> | Array<Cas3BookingSearchResult>,
+): results is Array<Cas3BookingSearchResult> =>
+  Boolean(results.length === 0 || (results[0] as Cas3BookingSearchResult).bedspace?.reference)
+
+export const bookingSearchResultsToCas3BookingSearchResults = (
+  results: Array<BookingSearchResult> | Array<Cas3BookingSearchResult>,
+): Array<Cas3BookingSearchResult> =>
+  isCas3BookingSearchResults(results)
+    ? results
+    : results.map(result => ({
+        bedspace: {
+          id: result.bed.id,
+          reference: result.bed.name,
+        },
+        booking: {
+          ...result.booking,
+          status: result.booking.status as Cas3BookingStatus,
+        },
+        person: result.person,
+        premises: result.premises,
+      }))
